flask-app/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-keycloak==4.1.0
flask-jwt-extended==4.6.0